feat: implement status changes
